+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Not released
+
+- Add: Option to ignore files
+
 ## v0.1.1 (2021-01-25)
 
 - Reduce Joplin min Version to 1.6.2